backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
# Banco de dados
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
# Variáveis de ambiente
python-dotenv==1.0.0
# Cache e sessões
redis==5.0.1
# Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Utilitários
python-multipart==0.0.6 